+++ django-data/image/biosample/views.py
@@ -1,5 +1,7 @@
 
 import os
+import re
+import json
 import redis
 import logging
 import traceback
@@ -35,7 +37,7 @@ from .tasks import SplitSubmissionTask
 logger = logging.getLogger(__name__)
 
 # while near
-TOKEN_DURATION_THRESHOLD = 3600
+TOKEN_DURATION_THRESHOLD = 3600*4
 
 # define a decouple config object
 settings_dir = os.path.join(settings.BASE_DIR, 'image')
@@ -162,15 +164,23 @@ class MyFormMixin(object):
             # logger exception. With repr() the exception name is rendered
             logger.error(repr(e))
 
-            # maybe I typed a wrong password or there is an issue in biosample
-            # log error in message and return form_invalid
-            # HINT: deal with two conditions?
+            # try to detect the error message
+            pattern = re.compile(r"(\{.*\})")
+            match = re.search(pattern, str(e))
 
-            # parse error message
-            messages.error(
-                self.request,
-                "Unable to generate token: %s" % str(e),
-                extra_tags="alert alert-dismissible alert-danger")
+            if match is None:
+                # parse error message
+                messages.error(
+                    self.request,
+                    "Unable to generate token: %s" % str(e),
+                    extra_tags="alert alert-dismissible alert-danger")
+
+            else:
+                data = json.loads(match.groups()[0])
+                messages.error(
+                    self.request,
+                    "Unable to generate token: %s" % data['message'],
+                    extra_tags="alert alert-dismissible alert-danger")
 
             # cant't return form_invalid here, since i need to process auth
             return None
@@ -299,7 +309,7 @@ class CreateUserView(LoginRequiredMixin, RegisterMixin, MyFormMixin, FormView):
 
     template_name = 'biosample/create_user.html'
     form_class = CreateUserForm
-    success_url_message = "Account created"
+    success_url_message = "AAP Account created"
 
     def deal_with_errors(self, error_message, exception):
         """Add messages to view for encountered errors